Refactor minor items in the project 'product list poc', improve its documentation and remove duplicate shadcn ui components
