+++ src/features/product-list/config/products.ts
